refac: Move entire dbus framework into platform folder.
The whole concept of DBus is platform specific.
