+++ src/api/auth/auth.cpp
@@ -19,7 +19,7 @@
 #include <vector>
 
 #include <common/common.hpp>
-#include <common/dbus.hpp>
+#include <common/platform/dbus.hpp>
 #include <common/error.hpp>
 #include <common/expected.hpp>
 #include <common/log.hpp>
