+++ src/common/platform/dbus/asio_libdbus/dbus.cpp
@@ -12,7 +12,7 @@
 //    See the License for the specific language governing permissions and
 //    limitations under the License.
 
-#include <common/dbus.hpp>
+#include <common/platform/dbus.hpp>
 
 #include <cassert>
 #include <functional>
